Conflicts:
	src/main.c

	new file:   CharacterDesignGuide.txt
	modified:   VisualC/Furnace/VisualC2010.vcxproj
	modified:   VisualC/Furnace/VisualC2010.vcxproj.filters
	new file:   images/Sofia.png
	new file:   include/characters.h
	modified:   include/gameObject.h
	modified:   include/player.h
	modified:   include/vectors.h
	modified:   src/game.c
	modified:   src/gameObject.c
	modified:   src/handleEvents.c
	modified:   src/main.c
	modified:   src/player.c
	modified:   src/textureManager.c
	modified:   src/vectors.c

+++ include/vectors.h
@@ -4,29 +4,29 @@
 #include <math.h>
  
 typedef struct {
-	float x;
-	float y;
+	int x;
+	int y;
 } vector2D;
 
 typedef struct {
-	float x;
-	float y;
-	float z;
+	int x;
+	int y;
+	int z;
 } vector3D;
 
 typedef struct {
-	float x;
-	float y;
-	float z;
-	float w;
+	int x;
+	int y;
+	int z;
+	int w;
 } vector4D;
 
-float FUR_lengthVector2D( vector2D v);
+int FUR_lengthVector2D( vector2D v);
 void FUR_norVector2D( vector2D *v );
 vector2D FUR_addVector2D( vector2D v1, vector2D v2 );
 vector2D FUR_subVector2D( vector2D v1, vector2D v2 );
-vector2D FUR_mulVector2D( vector2D v1, float scalar );
-vector2D FUR_divVector2D( vector2D v1, float scalar );
+vector2D FUR_mulVector2D( vector2D v1, int scalar );
+vector2D FUR_divVector2D( vector2D v1, int scalar );
 
 
 #endif // __vector2d__
